[full-ci] [tests-only] Do not run acceptance tests if only unit tests files are changed

+++ tests/drone/build-ocis-web.sh
@@ -2,7 +2,11 @@
 # Input parameters
 # $1 base directory
 
-cd /srv/app/src/github.com/owncloud/ocis || exit
-cd web || exit
-make build
-cp bin/web "$1"/ocis-web
+if test -f runUnitTestsOnly
+then echo 'Bye!'
+else
+	cd /srv/app/src/github.com/owncloud/ocis || exit
+	cd web || exit
+	make build
+	cp bin/web "$1"/ocis-web
+fi
