+++ tests/drone/build-web-app.sh
@@ -2,7 +2,11 @@
 # Input parameters
 # $1 the directory where the installed web app code is found
 
-yarn build
-mkdir -p /srv/config
-cp -r "$1"/tests/drone /srv/config
-ls -la /srv/config/drone
+if test -f runUnitTestsOnly
+then echo 'Bye!'
+else
+	yarn build
+	mkdir -p /srv/config
+	cp -r "$1"/tests/drone /srv/config
+	ls -la /srv/config/drone
+fi
